backend/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
fastapi==0.109.2
uvicorn[standard]>=0.20.0
lxml>=4.9.0
python-multipart==0.0.6
pydantic>=2.0.0
requests==2.31.0
imapclient>=3.0.0
PyJWT>=2.8.0
schedule>=1.2.0 